tests/wpt/web-platform-tests/css/css-break/parsing/break-inside-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fragmentation Module Level 3: parsing break-inside with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-break/#propdef-break-inside">
<meta name="assert" content="break-inside supports the full grammar 'auto | avoid | avoid-page | avoid-column | avoid-region'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("break-inside", "auto");
test_valid_value("break-inside", "avoid");
test_valid_value("break-inside", "avoid-page");
test_valid_value("break-inside", "avoid-column");
test_valid_value("break-inside", "avoid-region");
</script>
</body>
</html>
